- exceptions in concrete measures todo: logging warning situations and unittests for them

+++ ltr/measures/reciprocal_rank.h
@@ -42,7 +42,7 @@ namespace ltr {
      * Some decreasing function, usually f(x) = 1/x
      * @param pos - position of the first relevant object
      */
-    static double RRFormula(double pos);
+    static double RRFormula(int pos);
   };
 };
